refactor(ir): Isolate loop-invariant mat residency
