+++ docs/en/dev/passes/16-infer_tile_memory_space.md
@@ -94,9 +94,11 @@ A full `IRMutator` rewrite that produces the new function body:
 4. **Retargetable producer kwarg rewrite (`VisitStmt_(AssignStmt)`)** — for ops registered with `HasRetargetableMemoryKwarg()`, if Phase 1 resolved the output to a different space than the kwarg said (or the kwarg was absent), rewrite the `Call`'s `target_memory` kwarg and the result `TileType` to match. This keeps codegen and the assigned `Var` annotation in sync, and is necessary because Phase 1 may have resolved the producer using backward demand that the kwarg never saw.
 5. **LHS / RHS type sync** — when `VisitExpr_(Call)` rebuilds a `Call` via `OpRegistry` after argument substitution, the deduced result type may differ from the LHS `Var`'s original type (the rebuilt call sees inputs with new layouts). The mutator syncs the LHS Var's `TileType` to the rebuilt call's shape / dtype / memref / view while preserving the `memory_space_` chosen by Var rewrite, so roundtrip equality is preserved.
 
-### Phase 4 — Loop-invariant Mat residency (`LoopInvariantTileLoadHoister`)
+### Phase 4 — Loop-invariant Mat residency (`loop_invariant_mat_residency`)
 
-After all spaces are explicit, the pass recognizes an invariant prefix of the form `tile.load(GM → Mat) → tile.transpose_view* → tile.move/tile.extract(Mat → Left/Right)` and moves that prefix to the loop preheader. The chain must have one terminal `Left` / `Right` value used in the matching operand slot of exactly one `tile.matmul`, `tile.matmul_bias`, or `tile.matmul_acc`. This keeps the optimization specific to stationary matmul operands rather than turning it into general tile LICM. A stationary tensor-level matmul operand is therefore loaded once while a loop-dependent peer remains inside the loop and streams normally.
+After all spaces are explicit, a focused internal transform recognizes an invariant prefix of the form `tile.load(GM → Mat) → tile.transpose_view* → tile.move/tile.extract(Mat → Left/Right)` and moves that prefix to the loop preheader. The chain must have one terminal `Left` / `Right` value used in the matching operand slot of exactly one `tile.matmul`, `tile.matmul_bias`, or `tile.matmul_acc`. This keeps the optimization specific to stationary matmul operands rather than turning it into general tile LICM. For this supported single-use shape, a stationary tensor-level matmul operand is loaded once while a loop-dependent peer remains inside the loop and streams normally.
+
+This is a conservative first subset of the broader residency behavior requested by issue #2077, not a general tensor-level residency contract. A direct or externally entered InCore function has no caller evidence and therefore declines. When `AutoTileMatmulL0` fans one Mat panel out to multiple K-dependent L0 extracts, the current single-use chain recognizer also declines; the whole-panel GM→Mat load remains inside the outer user loop even though the K-dependent extracts remain correctly inside the inner K loop. Supporting that case requires analyzing GM→Mat panel residency independently from optional Mat→L0 prefix motion. The current caller filter also distinguishes canonical IR roots, but PyPTO does not yet attach a runtime `noalias` contract to distinct external tensor parameters; resolving that contract is required before this draft optimization is merge-ready.
 
 Eligibility begins with private compiler provenance. `ConvertTensorToTileOps` marks every generated `GM → Mat` bridge load; the marker survives printing, flattening, and L0 auto-tiling until this phase consumes it. Phase 4 then proves that a marked load participates in the exact stationary matmul-operand chain described above. A user-authored `tile.load(..., target_memory=Mat)` is not marked and is never hoisted by this optimization. This boundary keeps explicit tile programs under user control.
 
@@ -106,7 +108,7 @@ The initial legality rules are intentionally strict:
 - every moved assignment is an unconditional top-level loop-body statement;
 - the GM source is a direct `ParamDirection::In` tensor parameter on a compiler-marked Mat bridge load;
 - the InCore function has at least one direct call site in a root orchestration function (an orchestration function with no in-program caller), and every call site to that InCore function is such a direct root-orchestration call;
-- at every direct root-orchestration call site, the candidate `Tensor In` actual root and every writable `Tensor Out` / `Tensor InOut` actual root are known and the candidate root is distinct from every writable root; the InCore function also does not write that root locally, while unrelated scalars and peer read-only `Tensor In` roots do not participate in this proof;
+- at every direct root-orchestration call site, the candidate `Tensor In` actual root and every writable `Tensor Out` / `Tensor InOut` actual root are known and the candidate root is syntactically distinct from every writable root; the InCore function also does not write that root locally, while unrelated scalars and peer read-only `Tensor In` roots do not participate in this filter;
 - offsets, shapes, and the complete moved dependency prefix are loop invariant;
 - a direct control-flow or effect statement before the candidate closes the hoistable prefix;
 - no moved result is loop-carried or yielded;
@@ -138,7 +140,7 @@ for n, (acc,) in pl.range(0, 256, 128, init_values=(out,)):
     result = pl.yield_(pl.tile.store(c_n, [0, n], acc))
 ```
 
-The internal provenance attribute is omitted above for readability. The transformation requires a direct call from a root orchestration function that proves the candidate `lhs` root is distinct from the writable `out` root. The peer read-only `rhs` root is irrelevant to that proof. Without the required direct caller evidence, the original loop-local placement is retained.
+The internal provenance attribute is omitted above for readability. The transformation currently requires a direct call from a root orchestration function where the candidate `lhs` root is syntactically distinct from the writable `out` root. The peer read-only `rhs` root is irrelevant to that filter. Distinct IR roots alone do not prove that separately supplied runtime buffers cannot alias, so this caller condition remains an open draft-safety item. Without the required direct caller evidence, the original loop-local placement is retained.
 
 ## General memory-space inference example
